requirements.txt
asgiref==3.5.0
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
psycopg2-binary==2.8.6
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
Unidecode==1.3.4
